Here's the file for this week and last week
The file for this week runs with a three dimensional lattice topology
(because I'm curious whether the common result that 3D and higher
systems behave differently than 1D or 2D ones holds here), players who
are defectors with some probability and otherwise imitators of the kind
we saw last week, and a random selection rule for deciding where the
defectors get placed (and before the law of large numbers kicks in, how
many there are), and who moves and updates each round. It may need
different/more flexible rules eventually. It also doesn't have the
flashy visual effects Rick's did, and I'm curious about how to add them
(the output's hard to interpret in the command window).